% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFertilizerData.R
\name{getFertilizerData}
\alias{getFertilizerData}
\title{Get Fertilizer information}
\usage{
getFertilizerData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get Fertilizer information
}
\details{
x class SQL
}
\examples{
a value x
}
\keyword{MySQL}

