% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFieldData.R
\name{getFielddata}
\alias{getFielddata}
\title{Get field information}
\usage{
getFielddata(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Field information table (dataframe)
}
\description{
Get field information
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

