% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInjuriesData.R
\name{getInjuriesData}
\alias{getInjuriesData}
\title{Get SYT SKEP injury data}
\usage{
getInjuriesData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Injury data information table (dataframe)
}
\description{
This function queries the database for injury data
}
\details{
This function returns a database of injury data
}

