% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPestManagementData.R
\name{getPestManagementData}
\alias{getPestManagementData}
\title{Get SYT SKEP pest management data}
\usage{
getPestManagementData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Pest management information table (dataframe)
}
\description{
This function queries the database for pest management data
}
\details{
This function returns a database of pest management data
}

