% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemicData.R
\name{getSystemicData}
\alias{getSystemicData}
\title{Get SYT SKEP systemic injury data}
\usage{
getSystemicData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Systemic injury information table (dataframe)
}
\description{
This function queries the database for systemic injury data
}
\details{
This function returns a database of systemic injury data
}

