% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeedratingData.R
\name{getWeedratingData}
\alias{getWeedratingData}
\title{Get SYT SKEP weed rating data}
\usage{
getWeedratingData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Weed injury information table (dataframe)
}
\description{
This function queries the database for weed rating data
}
\details{
This function returns a database of weed injury rating data
}

