% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getYieldData.R
\name{getYieldData}
\alias{getYieldData}
\title{Get SYT SKEP yield data}
\usage{
getYieldData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Crop cut yield information table (dataframe)
}
\description{
This function queries the database for crop cut yield data
}
\details{
This function returns a database of crop cut yield data
}

