% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/SKEP.R
\name{loginSKEP}
\alias{loginSKEP}
\title{Assess to the data base}
\usage{
loginSKEP(username, password)
}
\arguments{
\item{username}{MUST BE Character}

\item{password}{MUST BE Character}
}
\value{
If two vectors, \code{x} and \code{y}, are provided, the output is a
  matrix with two columns, with the quantile normalized versions of
  \code{x} and \code{y}.
  If \code{y} is missing, \code{x} should be a matrix, in which case the
  output is a matrix of the same dimensions with the columns quantile
  normalized with respect to each other.
}
\description{
Assess to the data base
}
\details{
We access through the SKEP data base and sort out the data foolowing your param
}
\examples{
a value x
}
\keyword{MySQL}

