% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFertilizerData.R
\name{getFertilizerData}
\alias{getFertilizerData}
\title{Fertilizer information table (dataframe)}
\usage{
getFertilizerData(x)
}
\value{
dataframe
}
\description{
Fertilizer information table (dataframe)
}
\details{
x class SQL
}
\examples{
getFertilizerData(db) # not run
}
\keyword{MySQL}

