% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFertilizerData.R
\name{getFertilizerData}
\alias{getFertilizerData}
\title{Get SYT SKEP fertilizer data}
\usage{
getFertilizerData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Fertilizer information table (dataframe)
}
\description{
This function queries the database for fertilizer data
}
\details{
This function returns a database of fertilizer data
}

