% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFieldData.R
\name{getFielddata}
\alias{getFielddata}
\title{Get SYT SKEP field data}
\usage{
getFielddata(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Field information table (dataframe)
}
\description{
This function queries the database for field data
}
\details{
This function returns a database of field data
}

