% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getFieldData.R
\name{getFielddata}
\alias{getFielddata}
\title{getFertilizerdata: inquery fertilizer data from SKEP database}
\usage{
getFielddata(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Field information table (dataframe)
}
\description{
getFertilizerdata: inquery fertilizer data from SKEP database
}
\examples{
getFielddata(db) # not run
}
\keyword{MySQL}

