% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getFieldData.R
\name{getfielddata}
\alias{getfielddata}
\title{Assess to the data base}
\usage{
getfielddata(x)
}
\arguments{
\item{x}{is mySQL data}

\item{country}{is code}
}
\value{
Field information table (dataframe)
}
\description{
Assess to the data base
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

