% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getInjuriesData.R
\name{getInjuriesData}
\alias{getInjuriesData}
\title{Get Injury data}
\usage{
getInjuriesData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get Injury data
}
\details{

}
\examples{
a value x
}
\keyword{MySQL}

