% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPestManagementData.R
\name{getPestManagementData}
\alias{getPestManagementData}
\title{Get pest management data}
\usage{
getPestManagementData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get pest management data
}
\details{
x class SQL
}
\examples{
x value
}
\keyword{MySQL}

