% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSystemicData.R
\name{getSystemicData}
\alias{getSystemicData}
\title{Get systemic injury data}
\usage{
getSystemicData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get systemic injury data
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

