% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getSystemicData.R
\name{getSystemicData}
\alias{getSystemicData}
\title{Get Systemic injuruies data}
\usage{
getSystemicData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get Systemic injuruies data
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

