% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeedData.R
\name{getWeedData}
\alias{getWeedData}
\title{Get SYT SKEP weed data}
\usage{
getWeedData(x)
}
\arguments{
\item{x}{is the MySQL database hosted on Amazon Web Services}
}
\value{
Weed information table (dataframe)
}
\description{
This function queries the database for weed data
}
\details{
This function returns a database of weed data
}

