% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getWeedData.R
\name{getWeedData}
\alias{getWeedData}
\title{Assess to the data base}
\usage{
getWeedData(x)
}
\arguments{
\item{x}{is mySQL data}

\item{country}{is code}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Assess to the data base
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

