% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getWeedratingData.R
\name{getWeedratingData}
\alias{getWeedratingData}
\title{Get weed infestation data}
\usage{
getWeedratingData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Fertilizer information table (dataframe)
}
\description{
Get weed infestation data
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

