% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getYieldData.R
\name{getYieldData}
\alias{getYieldData}
\title{Get yield information data}
\usage{
getYieldData(x)
}
\arguments{
\item{x}{is mySQL data}
}
\value{
Farm infomation and yields information table (dataframe)
}
\description{
Get yield information data
}
\details{
x
}
\examples{
a value x
}
\keyword{MySQL}

