% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginSKEP.R
\name{loginSKEP}
\alias{loginSKEP}
\title{login to the SKEP data base}
\usage{
loginSKEP(username, password)
}
\arguments{
\item{username}{MUST BE Character}

\item{password}{MUST BE Character}
}
\description{
This function provides login capabilities to log into and
retrieve data from the IRRI/SYT SKEP Phase II database hosted on AWS.
}
\details{
This function provides access to the SYT SKEP database to retrieve
and organize data for analysis.
}
\examples{
\dontrun{
loginSKEP("username", "userpass")
}

}

