% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/loginSKEP.R
\name{loginSKEP}
\alias{loginSKEP}
\title{Assess to the data base}
\usage{
loginSKEP(username, password)
}
\arguments{
\item{username}{MUST BE Character}

\item{password}{MUST BE Character}
}
\value{
Data frame with Farm information
}
\description{
Assess to the data base
}
\details{
We access through the SKEP data base and sort out the data foolowing your param
}
\examples{
a value x
}
\keyword{MySQL}

