% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/loginSKEP.R
\name{loginSKEP}
\alias{loginSKEP}
\title{Provide login credentials for SYT SKEP database}
\usage{
loginSKEP(username, password)
}
\arguments{
\item{username}{MUST BE Character}

\item{password}{MUST BE Character}
}
\value{
Data frame with Farm information
Data frame with SYT SKEP data for further analysis
}
\description{
Provide login credentials for SYT SKEP database
}
\details{
Provides access to the SYT SKEP database to retrieve and organize data for analysis
}
\examples{
loginSKEP("username", "userpass") # not run
}
\keyword{MySQL}

